change int to uint64 or uint32

+++ node_jvm_stats.go
@@ -8,53 +8,53 @@ import (
 type NodeJVMStatsInfo struct {
 	JVM struct {
 		Threads struct {
-			Count     int `json:"count"`
-			PeakCount int `json:"peak_count"`
+			Count     uint64 `json:"count"`
+			PeakCount uint64 `json:"peak_count"`
 		} `json:"threads"`
 		Mem struct {
-			HeapUsedPercent         int `json:"heap_used_percent"`
-			HeapCommittedInBytes    int `json:"heap_committed_in_bytes"`
-			HeapMaxInBytes          int `json:"heap_max_in_bytes"`
-			HeapUsedInBytes         int `json:"heap_used_in_bytes"`
-			NonHeapUsedInBytes      int `json:"non_heap_used_in_bytes"`
-			NonHeapCommittedInBytes int `json:"non_heap_committed_in_bytes"`
+			HeapUsedPercent         uint64 `json:"heap_used_percent"`
+			HeapCommittedInBytes    uint64 `json:"heap_committed_in_bytes"`
+			HeapMaxInBytes          uint64 `json:"heap_max_in_bytes"`
+			HeapUsedInBytes         uint64 `json:"heap_used_in_bytes"`
+			NonHeapUsedInBytes      uint64 `json:"non_heap_used_in_bytes"`
+			NonHeapCommittedInBytes uint64 `json:"non_heap_committed_in_bytes"`
 			Pools                   struct {
 				Survivor struct {
-					PeakUsedInBytes  int `json:"peak_used_in_bytes"`
-					UsedInBytes      int `json:"used_in_bytes"`
-					PeakMaxInBytes   int `json:"peak_max_in_bytes"`
-					MaxInBytes       int `json:"max_in_bytes"`
-					CommittedInBytes int `json:"committed_in_bytes"`
+					PeakUsedInBytes  uint64 `json:"peak_used_in_bytes"`
+					UsedInBytes      uint64 `json:"used_in_bytes"`
+					PeakMaxInBytes   uint64 `json:"peak_max_in_bytes"`
+					MaxInBytes       uint64 `json:"max_in_bytes"`
+					CommittedInBytes uint64 `json:"committed_in_bytes"`
 				} `json:"survivor"`
 				Old struct {
-					PeakUsedInBytes  int `json:"peak_used_in_bytes"`
-					UsedInBytes      int `json:"used_in_bytes"`
-					PeakMaxInBytes   int `json:"peak_max_in_bytes"`
-					MaxInBytes       int `json:"max_in_bytes"`
-					CommittedInBytes int `json:"committed_in_bytes"`
+					PeakUsedInBytes  uint64 `json:"peak_used_in_bytes"`
+					UsedInBytes      uint64 `json:"used_in_bytes"`
+					PeakMaxInBytes   uint64 `json:"peak_max_in_bytes"`
+					MaxInBytes       uint64 `json:"max_in_bytes"`
+					CommittedInBytes uint64 `json:"committed_in_bytes"`
 				} `json:"old"`
 				Young struct {
-					PeakUsedInBytes  int `json:"peak_used_in_bytes"`
-					UsedInBytes      int `json:"used_in_bytes"`
-					PeakMaxInBytes   int `json:"peak_max_in_bytes"`
-					MaxInBytes       int `json:"max_in_bytes"`
-					CommittedInBytes int `json:"committed_in_bytes"`
+					PeakUsedInBytes  uint64 `json:"peak_used_in_bytes"`
+					UsedInBytes      uint64 `json:"used_in_bytes"`
+					PeakMaxInBytes   uint64 `json:"peak_max_in_bytes"`
+					MaxInBytes       uint64 `json:"max_in_bytes"`
+					CommittedInBytes uint64 `json:"committed_in_bytes"`
 				} `json:"young"`
 			} `json:"pools"`
 		} `json:"mem"`
 		GC struct {
 			Collectors struct {
 				Old struct {
-					CollectionTimeInMillis int `json:"collection_time_in_millis"`
-					CollectionCount        int `json:"collection_count"`
+					CollectionTimeInMillis uint64 `json:"collection_time_in_millis"`
+					CollectionCount        uint64 `json:"collection_count"`
 				} `json:"old"`
 				Young struct {
-					CollectionTimeInMillis int `json:"collection_time_in_millis"`
-					CollectionCount        int `json:"collection_count"`
+					CollectionTimeInMillis uint64 `json:"collection_time_in_millis"`
+					CollectionCount        uint64 `json:"collection_count"`
 				} `json:"young"`
 			} `json:"collectors"`
 		} `json:"gc"`
-		UptimeInMillis int `json:"uptime_in_millis"`
+		UptimeInMillis uint64 `json:"uptime_in_millis"`
 	} `json:"jvm"`
 }
